+++ framework/src/main/java/org/checkerframework/common/basetype/BaseTypeVisitor.java
@@ -2194,7 +2194,7 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
      * @param tree method or constructor invocation tree
      */
     protected void checkVarargs(AnnotatedExecutableType invokedMethod, Tree tree) {
-        if (!TreeUtils.isVarArgs(tree)) {
+        if (!TreeUtils.isVarargsCall(tree)) {
             // If not a varargs invocation, type checking is already done in checkArguments.
             return;
         }
@@ -3970,19 +3970,22 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
             List<? extends ExpressionTree> passedArgs,
             CharSequence executableName,
             List<?> paramNames) {
-        int size = requiredTypes.size();
-        assert size == passedArgs.size()
-                : "size mismatch between required args ("
-                        + requiredTypes
-                        + ") and passed args ("
-                        + passedArgs
-                        + ")";
+        int numRequired = requiredTypes.size();
+        assert numRequired == passedArgs.size()
+                : String.format(
+                        "numRequired %d should equal %d in checkArguments(%s, %s, %s, %s)",
+                        numRequired,
+                        passedArgs.size(),
+                        listToString(requiredTypes),
+                        listToString(passedArgs),
+                        executableName,
+                        listToString(paramNames));
         int maxParamNamesIndex = paramNames.size() - 1;
         // Rather weak assertion, due to how varargs parameters are treated.
-        assert size >= maxParamNamesIndex
+        assert numRequired >= maxParamNamesIndex
                 : String.format(
                         "mismatched lengths %d %d %d checkArguments(%s, %s, %s, %s)",
-                        size,
+                        numRequired,
                         passedArgs.size(),
                         paramNames.size(),
                         listToString(requiredTypes),
@@ -3990,7 +3993,7 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
                         executableName,
                         listToString(paramNames));
 
-        for (int i = 0; i < size; ++i) {
+        for (int i = 0; i < numRequired; ++i) {
             AnnotatedTypeMirror requiredType = requiredTypes.get(i);
             ExpressionTree passedArg = passedArgs.get(i);
             Object paramName = paramNames.get(Math.min(i, maxParamNamesIndex));