+++ framework/src/main/java/org/checkerframework/framework/util/AnnotatedTypes.java
@@ -29,7 +29,6 @@ import org.checkerframework.javacutil.AnnotationMirrorSet;
 import org.checkerframework.javacutil.AnnotationUtils;
 import org.checkerframework.javacutil.BugInCF;
 import org.checkerframework.javacutil.ElementUtils;
-import org.checkerframework.javacutil.SystemUtil;
 import org.checkerframework.javacutil.TreeUtils;
 import org.checkerframework.javacutil.TypesUtils;
 import org.plumelib.util.CollectionsPlume;
@@ -51,6 +50,7 @@ import java.util.Set;
 
 import javax.lang.model.element.AnnotationMirror;
 import javax.lang.model.element.Element;
+import javax.lang.model.element.ElementKind;
 import javax.lang.model.element.ExecutableElement;
 import javax.lang.model.element.TypeElement;
 import javax.lang.model.element.TypeParameterElement;
@@ -1014,6 +1014,31 @@ public class AnnotatedTypes {
     }
 
     /**
+     * Returns the method parameters for the invoked method, with the same number of arguments
+     * passed in the methodInvocation tree.
+     *
+     * <p>If the invoked method is not a vararg method or it is a vararg method but the invocation
+     * passes an array to the vararg parameter, it would simply return the method parameters.
+     *
+     * <p>Otherwise, it would return the list of parameters as if the vararg is expanded to match
+     * the size of the passed arguments.
+     *
+     * @param atypeFactory the type factory to use for fetching annotated types
+     * @param method the method's type
+     * @param args the arguments to the method invocation
+     * @return the types that the method invocation arguments need to be subtype of
+     * @deprecated Use {@link #adaptParameters(AnnotatedTypeFactory, AnnotatedExecutableType, List,
+     *     Tree)} instead
+     */
+    @Deprecated // 2022-04-21
+    public static List<AnnotatedTypeMirror> expandVarArgsParameters(
+            AnnotatedTypeFactory atypeFactory,
+            AnnotatedExecutableType method,
+            List<? extends ExpressionTree> args) {
+        return adaptParameters(atypeFactory, method, args, null);
+    }
+
+    /**
      * Returns the method parameters for the invoked method (or constructor), with the same number
      * of arguments as passed to the invocation tree.
      *
@@ -1024,7 +1049,7 @@ public class AnnotatedTypes {
      * @param atypeFactory the type factory to use for fetching annotated types
      * @param method the method or constructor's type
      * @param args the arguments to the method or constructor invocation
-     * @param tree the NewClassTree if method is a constructor
+     * @param invok the method or constructor invocation
      * @return a list of the types that the invocation arguments need to be subtype of; has the same
      *     length as {@code args}
      */
@@ -1032,32 +1057,48 @@ public class AnnotatedTypes {
             AnnotatedTypeFactory atypeFactory,
             AnnotatedExecutableType method,
             List<? extends ExpressionTree> args,
-            @Nullable NewClassTree tree) {
+            Tree invok) {
+
         List<AnnotatedTypeMirror> parameters = method.getParameterTypes();
-        // Handle anonymous constructors that extend a class with an enclosing type.
-        // There is a mismatch between the number of parameters and arguments when
-        // the following conditions are met:
-        // 1. Java version >= 11
-        // 2. the method is an anonymous constructor
-        // 3. the constructor is invoked with an explicit enclosing expression
-        // In the case, we should remove the first parameter.
-        if (SystemUtil.jreVersion >= 11
-                && tree != null
-                && TreeUtils.isAnonymousConstructorWithExplicitEnclosingExpression(
-                        method.getElement(), tree)) {
-            if (parameters.size() != args.size() || args.isEmpty()) {
-                List<AnnotatedTypeMirror> p = new ArrayList<>(parameters.size());
-                p.addAll(parameters.subList(1, parameters.size()));
-                parameters = p;
+
+        // Handle anonymous constructors that extend a class with an enclosing type,
+        // as in `new MyClass(){ ... }`.
+        if (method.getElement().getKind() == ElementKind.CONSTRUCTOR
+                && method.getElement().getEnclosingElement().getSimpleName().contentEquals("")) {
+            DeclaredType t =
+                    TypesUtils.getSuperClassOrInterface(
+                            method.getElement().getEnclosingElement().asType(), atypeFactory.types);
+            if (t.getEnclosingType() != null) {
+                if (!parameters.isEmpty()) {
+                    if (atypeFactory.types.isSameType(
+                            t.getEnclosingType(), parameters.get(0).getUnderlyingType())) {
+                        if (args.isEmpty()
+                                || !atypeFactory.types.isSameType(
+                                        TreeUtils.typeOf(args.get(0)),
+                                        parameters.get(0).getUnderlyingType())) {
+                            parameters = parameters.subList(1, parameters.size());
+                        }
+                    }
+                }
             }
         }
 
         // Handle vararg methods.
-        if (!method.getElement().isVarArgs()) {
+        if (!TreeUtils.isVarargsCall(invok)) {
             return parameters;
         }
+        if (parameters.isEmpty()) {
+            throw new BugInCF("isVarargsCall but parameters is empty: %s", invok);
+        }
 
-        AnnotatedArrayType varargs = (AnnotatedArrayType) parameters.get(parameters.size() - 1);
+        AnnotatedTypeMirror lastParam = parameters.get(parameters.size() - 1);
+        if (!(lastParam instanceof AnnotatedArrayType)) {
+            throw new BugInCF(
+                    String.format(
+                            "for varargs call %s, last parameter %s is not an array",
+                            invok, lastParam));
+        }
+        AnnotatedArrayType varargs = (AnnotatedArrayType) lastParam;
 
         if (parameters.size() == args.size()) {
             // Check if one sent an element or an array