+++ framework/src/main/java/org/checkerframework/framework/util/dependenttypes/DependentTypesHelper.java
@@ -20,6 +20,7 @@ import org.checkerframework.dataflow.expression.FormalParameter;
 import org.checkerframework.dataflow.expression.JavaExpression;
 import org.checkerframework.dataflow.expression.JavaExpressionConverter;
 import org.checkerframework.dataflow.expression.LocalVariable;
+import org.checkerframework.dataflow.expression.SuperReference;
 import org.checkerframework.dataflow.expression.ThisReference;
 import org.checkerframework.dataflow.expression.Unknown;
 import org.checkerframework.framework.source.SourceChecker;
@@ -785,6 +786,12 @@ public class DependentTypesHelper {
                                 @Override
                                 public JavaExpression visitThisReference(
                                         ThisReference thisRef, Void unused) {
+                                    throw new FoundLocalVarException();
+                                }
+
+                                @Override
+                                public JavaExpression visitSuperReference(
+                                        SuperReference superRef, Void unused) {
                                     throw new FoundLocalVarException();
                                 }
                             };