+++ framework/src/test/java/org/checkerframework/framework/testchecker/javaexpression/qual/FEBottom.java
@@ -1,0 +1,10 @@
+package org.checkerframework.framework.testchecker.javaexpression.qual;
+
+import org.checkerframework.framework.qual.SubtypeOf;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Target;
+
+@SubtypeOf({FlowExp.class})
+@Target({ElementType.TYPE_USE, ElementType.TYPE_PARAMETER})
+public @interface FEBottom {}